+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.3.1"></a>
+## [0.3.1](https://github.com/Jam3/adviser/compare/v0.3.0...v0.3.1) (2019-08-23)
+
+### Features
+
+* add events to the engine
+* add loader UI to the terminal
+
+
 <a name="0.3.0"></a>
 # [0.3.0](https://github.com/Jam3/adviser/compare/v0.2.2...v0.3.0) (2019-08-13)